break the components of jit.cc out into more-manageable pieces
