+++ tile/targets/cpu/link_names.cc
@@ -1,0 +1,19 @@
+// Copyright 2019, IntelCorp.
+
+#include "tile/targets/cpu/link_names.h"
+
+namespace vertexai {
+namespace tile {
+namespace targets {
+namespace cpu {
+
+const char invoker_name_[] = "__invoke_";
+const char arena_name_[] = "__arena";
+const char profile_count_name_[] = "__profile_count_";
+const char profile_ticks_name_[] = "__profile_ticks_";
+const char profile_loop_body_name_[] = "__profile_loop_body_";
+
+}  // namespace cpu
+}  // namespace targets
+}  // namespace tile
+}  // namespace vertexai
