+++ tile/targets/cpu/link_names.h
@@ -1,0 +1,19 @@
+// Copyright 2019, Intel Corp.
+
+#pragma once
+
+namespace vertexai {
+namespace tile {
+namespace targets {
+namespace cpu {
+
+extern const char invoker_name_[];
+extern const char arena_name_[];
+extern const char profile_count_name_[];
+extern const char profile_ticks_name_[];
+extern const char profile_loop_body_name_[];
+
+}  // namespace cpu
+}  // namespace targets
+}  // namespace tile
+}  // namespace vertexai
